+++ Sources/MusicBarApp/AppEdition.swift
@@ -1,13 +1,4 @@
 enum AppEdition {
-    #if PREVIEW
-    static let name = "MusicBar Preview"
-
-    static let supportsAutoLyricsWindow = false
-    static let supportsLaunchAtLogin = false
-    static let supportsLyricsWindowPinning = false
-    static let supportsLyricsWindowOpacity = false
-    static let supportsLyricLineSeeking = false
-    #else
     static let name = "MusicBar"
 
     static let supportsAutoLyricsWindow = true
@@ -15,7 +6,8 @@ enum AppEdition {
     static let supportsLyricsWindowPinning = true
     static let supportsLyricsWindowOpacity = true
     static let supportsLyricLineSeeking = true
-    #endif
 
     static let githubURLString = "https://github.com/aronahhhh/MusicBar"
+    static let purchaseURLString = "https://github.com/aronahhhh/MusicBar/releases/latest"
+    static let trialDurationDays = 7
 }